third_party/WebKit/LayoutTests/inspector/sources/debugger/promise-tracker-expected.txt

Tests promise tracker in debugger.

Set timer for test function.
Promise (new):
    id: 0
    status: pending
    parent id: undefined
    testFunction promise-tracker.html:9
Promise (update):
    id: 0
    status: resolved
    parent id: undefined
    testFunction promise-tracker.html:9
Promise (new):
    id: 1
    status: pending
    parent id: undefined
    testFunction promise-tracker.html:13
Promise (update):
    id: 1
    status: pending
    parent id: 0
    testFunction promise-tracker.html:13
Promise (new):
    id: 2
    status: pending
    parent id: undefined
    testFunction promise-tracker.html:14
Promise (update):
    id: 2
    status: pending
    parent id: 1
    testFunction promise-tracker.html:14
Promise (update):
    id: 1
    status: resolved
    parent id: 0
    testFunction promise-tracker.html:13
Script execution paused.
Got promise by id: true
Resolved object class: Promise
Promise:
    status: resolved
    value: Resolved!
Script execution resumed.
